LONG MESSAGE
This commit is extensive and it goes against the 'commit often' rule.
1) The content is no longer managed through attributes,
but rather directly through post meta.
2) I decoupled some functionality through utils.js.
3) Set a state regarding data being updated through API call,
because it was interfering with the useEffect.
4) useEffect only runs when content is getting manually edited,
but not when programatically changed through makeApiCall method.
5) There was some changes in makeApiCall regarding the POST body,
particularly formatting the content to exclude the block itself
from the content and cleaning the markup that Gutenberg uses,
sending a string with no tags.
6) If the response from the API call is ok, it updates the meta value
and replace the inner blocks.
7) Finally, in the API call, the updatingData state is set.
8) This also implements InnerBlocks Component instead of RichText.
9) Some imports changed, some were deleted, and some added.